adding the sort function in the linkedlist.h
